refactor(package): C/Q 구분 삭제
